Fix loop in check_serial_port when connected via usbip
When I tried running the integration over usbip, I encountered a problem in check_serial_port method. It looks like the reader thread is reading too fast and it reads the writer's  data before it gets flushed to the alarm control unit, so the unit never gets the data and never responds. As a quick fix, I added a sleep to the reader loop to slow down reading.